Fixed comments in PR
