+++ components/FloatingActionButton/FloatingActionButton.jsx
@@ -1,32 +1,19 @@
 import styles from '@/styles/FloatingActionButton.module.css'
 import { Icon } from '@/components'
-import { useEffect, useState } from 'react'
+import { useScroll } from '@/hooks'
+import Link from 'next/link'
 
 const FloatingActionButton = () => {
-    const [show, setShow] = useState(false)
-
-    useEffect(() => {
-        const scrollListener = () => {
-            setShow(window.scrollY > 10)
-        }
-        window.addEventListener('scroll', scrollListener)
-
-        return () => {
-            window.removeEventListener('scroll', scrollListener)
-        }
-    }, [])
+    const show = useScroll(10)
 
     return (
-        <>
-            {show ? (
-                <div
-                    className={`${styles.fabWrapper} flex-row-center`}
-                    onClick={() => void window.scrollTo({ top: 0 })}
-                >
-                    <Icon width={50} height={50} variant='arrow-up' />
-                </div>
-            ) : null}
-        </>
+        show && (
+            <div className={`${styles.fabWrapper} flex-row-center`}>
+                <Link href='#'>
+                    <Icon width={50} height={50} variant='arrowUp'></Icon>
+                </Link>
+            </div>
+        )
     )
 }
