+++ components/Header/Nav.jsx
@@ -1,17 +1,19 @@
+import Link from 'next/link'
+
 const Nav = (props) => {
     const { className, closeModal } = props
 
     return (
         <div className={className}>
-            <a href='#about' className='margin' onClick={closeModal}>
+            <Link href='#about' className='margin' onClick={closeModal}>
                 About
-            </a>
-            <a href='#join-us' className='margin' onClick={closeModal}>
+            </Link>
+            <Link href='#join-us' className='margin' onClick={closeModal}>
                 Join us
-            </a>
-            <a href='#contact' className='margin' onClick={closeModal}>
+            </Link>
+            <Link href='#contact' className='margin' onClick={closeModal}>
                 Contact
-            </a>
+            </Link>
         </div>
     )
 }
