+++ components/Icon/Icon.jsx
@@ -3,15 +3,10 @@ import menu from '../../public/images/menu-icon.svg'
 import close from '../../public/images/close-icon.svg'
 import arrowUp from '../../public/images/arrow-up-icon.svg'
 
-const getSrc = (variant) => {
-    switch (variant) {
-        case 'menu':
-            return menu
-        case 'close':
-            return close
-        case 'arrow-up':
-            return arrowUp
-    }
+const icons = {
+    menu,
+    close,
+    arrowUp,
 }
 
 const Icon = (props) => {
@@ -20,6 +15,6 @@ const Icon = (props) => {
     if (width) size.width = width
     if (height) size.height = height
 
-    return <Image {...size} alt={`${variant} icon`} src={getSrc(variant)} />
+    return <Image {...size} alt={`${variant} icon`} src={icons[variant]} />
 }
 export default Icon
