+++ styles/FloatingActionButton.module.css
@@ -2,4 +2,5 @@
     position: fixed;
     bottom: 3rem;
     left: 50%;
+    transform: translateX(-50%);
 }
