+++ styles/globals.css
@@ -19,7 +19,7 @@
     --tertiary-background: rgba(194, 18, 18, 0.15);
 
     --neutral-black: #1a1a1a;
-    --neutral-dark: #525252;
+    --neutral-dark: #39393d;
     --neutral-medium-dark: #525252;
     --neutral-medium: #9b9b9f;
     --neutral-medium-light: #c7c8d0;
@@ -38,6 +38,10 @@
     --button-text-disabled: #39393d;
     --button-outline-disabled: #989799;
     --button-bg-outline-disabled: #f5f5f5;
+
+    --nav-text-color: #383f60;
+    --modal-backdrop: rgba(0, 0, 0, 0.5);
+    --menu-border: #ccd2eb;
 
     --base-font-size: 1rem;
     --spacing-unit: 16px;
